updated Sensor test
