+++ test_flask/Sensor_test.py
@@ -1,17 +1,21 @@
 from flask import Flask, render_template
 import BottleSensors as bs
 import os
+import bottle_counter as gbc
 
 app = Flask(__name__)
 
 @app.route("/")
 def index():
     sensorSts = bs.bottle_counter()
+    count     = gbc.sold_bottle_count
     templateData = {
         'title': 'GPIO input Status!',
         'button': sensorSts,
-    }
+        'quantity': count,
     
+    }
+                         #'Mate_website.html'
     return render_template('index.html', **templateData)
 
 def startFlask():
